requirements.txt
# Django Core
Django==4.2.7
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-filter==24.3

# API Documentation
drf-yasg==1.21.7

# Admin Interface
django-admin-interface==0.26.0

# CORS Headers
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8

# Production Server
gunicorn==21.2.0

# Utilities
pytz==2023.3
